requirements.txt
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
exceptiongroup==1.1.1
fastapi==0.96.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
pydantic==1.10.9
PyJWT==2.7.0
pymongo==4.3.3
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
